re-style order food card
