+++ order-system-ionic/src/components/FoodItemCards.tsx
@@ -1,6 +1,6 @@
 import CartFoodItemCard from "./FoodItemCards/CartFoodItemCard";
 import MenuFoodItemCard from "./FoodItemCards/MenuFoodItemCard/MenuFoodItemCard";
-import OrderFoodItemCard from "./FoodItemCards/OrderFoodItemCard";
+import OrderFoodItemCard from "./FoodItemCards/OrderFoodItemCard/OrderFoodItemCard";
 import SplitBillFoodItemCard from "./FoodItemCards/SplitBillFoodItemCard";
 
 export { MenuFoodItemCard, OrderFoodItemCard, CartFoodItemCard, SplitBillFoodItemCard };
