+++ order-system-ionic/src/components/FoodItemCards/OrderFoodItemCard/OrderFoodItemCard.module.scss
@@ -1,0 +1,40 @@
+.ionCardContainer {
+    border-radius: 1rem;
+    background: 0%;
+    box-shadow: none;
+}
+
+.divContent {
+    padding: 2px;
+    display: flex;
+    flex-direction: row;
+    cursor: pointer;
+}
+
+.cardImage {
+    width: 7.5rem;
+    height: 7.5rem;
+    object-fit: cover;
+    padding: 1rem;
+    border-radius: 25%;
+}
+
+.foodName {
+    font-size: 1.2rem;
+}
+
+.price {
+    position: absolute;
+    top: 1;
+    right: 0;
+    padding: 1rem;
+}
+
+.contentContainer {
+    padding-top: 15px;
+    flex: 2;
+    display: flex;
+    flex-direction: column;
+    padding-left: 0;
+    margin-left: 0;
+}
